Add product editing and creation with type, add admin page styles and other misc changes

+++ src/Controller/Admin/AdminController.php
@@ -2,7 +2,9 @@
 
 namespace App\Controller\Admin;
 
+use App\Entity\Product;
 use App\Form\DesignType;
+use App\Form\ProductType;
 use App\Repository\DesignRepository;
 use App\Repository\ProductRepository;
 use App\Service\ProductOrder;
@@ -15,6 +17,8 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/admin')]
 class AdminController extends AbstractController
 {
+    private const PAGE_LIMIT = 15;
+
     public function __construct(
         private EntityManagerInterface $em,
         private DesignRepository $designRepository,
@@ -32,17 +36,15 @@ class AdminController extends AbstractController
         ]);
     }
 
-    //TODO - kai kursi producto pridejima nepamirsk prisetinti jam orderi lygu jo id
-    // arba vienu didesniu nei paskutinis order value
-
     //TODO 2 - saugoti pasirinkta pre-made order value ir atnaujinti select lista su ja
     // taip pat jei bus custom rodyt tai (gal veliau save custom order?)
 
     #[Route('/products', name: 'admin_products')]
-    public function productPage()
+    public function productPage(Request $request)
     {
         return $this->render('admin/products.html.twig', [
-            'products' => $this->productRepository->getAllProductsWithOrder(),
+            'query' => $request->query->get('query'),
+            'products' => $this->productRepository->findProductsPaginatedWithSearch(self::PAGE_LIMIT, $request->query->get('query'), $request->query->get('page')),
             'design' => $this->designRepository->getDesignProductLimit()
         ]);
     }
@@ -57,14 +59,68 @@ class AdminController extends AbstractController
         return $this->redirectToRoute('admin_products');
     }
 
+    #[Route('/products/add', name: 'product_add')]
+    public function addProduct(Request $request)
+    {
+        $product = new Product();
+        $form = $this->createForm(ProductType::class, $product);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $lastOrderItem = $this->productRepository->findOneBy([], ['displayOrder' => 'DESC']);
+
+            $displayOrder = 1;
+            if ($lastOrderItem) {
+                $displayOrder = $lastOrderItem->getDisplayOrder()+1;
+            }
+
+            $product->setDisplayOrder($displayOrder);
+            $this->em->persist($product);
+            $this->em->flush();
+
+            $this->addFlash('success', 'Saved!');
+            return $this->redirectToRoute('admin_products');
+        }
+
+        return $this->render("admin/_form.html.twig", [
+            'action' => 'Adding a new product',
+            'form' => $form->createView()
+        ]);
+    }
+
+    #[Route('/products/edit/{id}', name: 'product_edit')]
+    public function editProduct(Request $request, int $id): Response
+    {
+        $product = $this->productRepository->findOneBy(['id' => $id]);
+
+        $form = $this->createForm(ProductType::class, $product);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->em->flush();
+
+            $this->addFlash('success', 'Edited successfully!');
+            return $this->redirectToRoute('admin_products');
+        }
+
+        return $this->render("admin/_form.html.twig", [
+            'action' => 'Editing product - ' . $product->getTitle(),
+            'form' => $form->createView()
+        ]);
+    }
+
     #[Route('/products/delete/{id}', name: 'product_delete')]
     public function deleteProduct(Request $request, int $id)
     {
         $product = $this->productRepository->findOneBy(['id' => $id]);
 
-        $this->em->remove($product);
-        $this->em->flush();
-        $this->addFlash('success', 'Product has been deleted');
+        if ($product) {
+            $this->em->remove($product);
+            $this->em->flush();
+            $this->addFlash('success', 'Product has been deleted');
+        } else {
+            $this->addFlash('warning', 'No such product found');
+        }
 
         $referer = $request->headers->get('referer');
         return $this->redirect($referer);
